+++ src/main/java/likelion/senifood/entity/Lunchbox.java
@@ -17,7 +17,7 @@ public class Lunchbox {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer lunchbox_id;
+    private Long lunchbox_id;
     private String lunchbox_title;
     private String lunchbox_imageURL;
     private String lunchbox_price;
